# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Mizuki API",
        "description": "API cho hệ thống thương mại mỹ phẩm và đặt lịch clinic Mizuki",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/technician/appointments": {
            "get": {
                "tags": [
                    "Technician Appointments"
                ],
                "summary": "Danh sách lịch hẹn được phân công",
                "operationId": "technicianListAppointments",
                "responses": {
                    "200": {
                        "description": "Danh sách lịch hẹn"
                    },
                    "403": {
                        "description": "Không có quyền"
                    }
                }
            }
        },
        "/api/v1/technician/appointments/{id}": {
            "get": {
                "tags": [
                    "Technician Appointments"
                ],
                "summary": "Chi tiết lịch hẹn được phân công",
                "operationId": "technicianShowAppointment",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Chi tiết lịch hẹn"
                    },
                    "404": {
                        "description": "Không tìm thấy"
                    }
                }
            }
        },
        "/api/v1/technician/appointments/{id}/start": {
            "post": {
                "tags": [
                    "Technician Appointments"
                ],
                "summary": "Bắt đầu lịch hẹn được phân công",
                "operationId": "technicianStartAppointment",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Đã bắt đầu"
                    },
                    "422": {
                        "description": "Trạng thái không hợp lệ"
                    }
                }
            }
        },
        "/api/v1/technician/appointments/{id}/complete": {
            "post": {
                "tags": [
                    "Technician Appointments"
                ],
                "summary": "Hoàn thành lịch hẹn được phân công",
                "operationId": "technicianCompleteAppointment",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Đã hoàn thành"
                    },
                    "422": {
                        "description": "Trạng thái không hợp lệ"
                    }
                }
            }
        }
    },
...
}